Vasija se renderiza
El problema estaba en que el buffer de indices era un GLubyte. Lo he tenido que cambiar a GLuint, y con eso la generación del buffer y el renderizado de los objetos. Ahora funciona correctamente

+++ Code/Sources/AssimpMesh.cpp
@@ -6,7 +6,7 @@ namespace udit
     AssimpMesh::AssimpMesh(const std::string& filepath)
     {
         loadModel(filepath);
-        generateBuffers(); // Llama al método heredado para subir a GPU
+        generateBuffers();
     }
 
     void AssimpMesh::loadModel(const std::string& path)
@@ -66,7 +66,7 @@ namespace udit
             aiFace face = mesh->mFaces[i];
             for (unsigned int j = 0; j < face.mNumIndices; ++j)
             {
-                indices.push_back(static_cast<GLubyte>(face.mIndices[j]));
+                indices.push_back(static_cast<GLuint>(face.mIndices[j]));
             }
         }
     }
